requirements.txt
# OSINT Peru 2026 - Dependencies
pandas>=2.0.0,<3.0.0
requests>=2.28.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0,<7.0
duckdb>=1.0.0,<2.0.0
pyarrow>=14.0.0,<18.0.0
loguru>=0.7.0,<1.0.0
eventregistry>=9.0,<10.0
